🩹 fix : DB 스키마, 쿼리 수정

+++ src/db/sql/user_coordinates.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS coordinates_backup
 (
     id			VARCHAR(36) PRIMARY KEY,
     device_id	VARCHAR(255) UNIQUE NOT NULL,
-    x			INT,
-    y			INT
+    x			FLOAT,
+    y			FLOAT
 );
